Add files via upload
Hey Guys!

The draft of the project proposal is done!
Please go through the report and lemme know if any changes are required. Also, just add your student ids and an appropriate title on the cover page.

PS: I selected a different dataset to provide us with more room to play for the final project. I am uploading it in this Git Repo as well :)